roll.sh: fix playing audio on GNU/Linux (aplay)
The argument -Dplug:default was stopping sound working; removing it makes it
work again.

Rather to my surprise, it also worked with the rest of the arguments
removed. This seems odd given that it claims to be downloading a raw audio
file, but removing -q showed that the audio data aplay was being fed is in
fact a WAV file, and downloading the file and checking it confirms this.

Hence, refer to “wav audio” rather than “raw audio”, and remove all
command-line arguments to aplay. Win!